removed password from userData response

+++ controllers/user.js
@@ -1,5 +1,6 @@
 import User from "../models/User.js";
 import { createError } from "../utils/errors.js";
+import { getUserData } from "../utils/getUserData.js";
 
 
 // create a user
@@ -7,7 +8,8 @@ export const createUser =  async (req, res, next) => {
   const newUser = new User(req.body);
   try{
     const savedUser = await newUser.save();
-    res.status(201).json(savedUser);
+    const userData = getUserData(savedUser);
+    res.status(201).json(userData);
   }catch(err){
     next(err);
   };
@@ -17,7 +19,8 @@ export const createUser =  async (req, res, next) => {
 export const getAllUsers =  async (req, res, next) => {
   try{
     const users = await User.find({});
-    res.status(200).json(users);
+    const usersData = users.map(user => getUserData(user));
+    res.status(200).json(usersData);
   }catch(err){
     next(err);
   }
@@ -27,7 +30,8 @@ export const getAllUsers =  async (req, res, next) => {
 export const getOwnProfile =  async (req, res, next) => {
   try{
     const user = await User.findById(req.user.id );
-    res.status(200).json(user);
+    const userData = getUserData(user);
+    res.status(200).json(userData);
   }catch(err){
     next(err);
   }
@@ -37,7 +41,8 @@ export const getOwnProfile =  async (req, res, next) => {
 export const getUser =  async (req, res, next) => {
   try{
     const user = await User.findById(req.params.userId);
-    res.status(200).json(user);
+    const userData = getUserData(user);
+    res.status(200).json(userData);
   }catch(err){
     next(err);
   }
@@ -52,7 +57,8 @@ export const updatedUser =  async (req, res, next) => {
     const updatedUser = await User.findByIdAndUpdate(req.params.userId, req.body, {
       new: true
     });
-    res.status(200).json(updatedUser);
+    const userData = getUserData(updatedUser);
+    res.status(200).json(userData);
   }catch(err){
     next(err);
   }
@@ -61,8 +67,12 @@ export const updatedUser =  async (req, res, next) => {
 // delete a user by id
 export const deleteUser = async (req, res, next) => {
   try{
+    if(req.params.userId !== req.user.id){
+      return next(createError({message: 'You are not authorized to delete this user', status: 403}));
+    }
     const deletedUser = await User.findByIdAndDelete(req.params.userId);
-    res.status(200).json(deletedUser);
+    const userData = getUserData(deletedUser);
+    res.status(200).json(userData);
   }catch(err){
     next(err);
   }
